# Conflicts:
#	health-services/individual/src/main/java/org/egov/individual/repository/IndividualRepository.java
#	health-services/individual/src/main/java/org/egov/individual/service/IndividualService.java
#	health-services/individual/src/test/java/org/egov/individual/service/IndividualEnrichmentServiceDeleteTest.java
#	health-services/individual/src/test/java/org/egov/individual/service/IndividualEnrichmentServiceUpdateTest.java
#	health-services/individual/src/test/java/org/egov/individual/web/controllers/IndividualApiControllerTest.java

+++ health-services/stock/src/main/java/org/egov/stock/repository/rowmapper/StockReconciliationRowMapper.java
@@ -3,8 +3,8 @@ package org.egov.stock.repository.rowmapper;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import digit.models.coremodels.AuditDetails;
-import org.egov.stock.web.models.AdditionalFields;
-import org.egov.stock.web.models.StockReconciliation;
+import org.egov.common.models.stock.AdditionalFields;
+import org.egov.common.models.stock.StockReconciliation;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
 
@@ -27,7 +27,7 @@ public class StockReconciliationRowMapper implements RowMapper<StockReconciliati
                     .id(resultSet.getString("id"))
                     .clientReferenceId(resultSet.getString("clientReferenceId"))
                     .tenantId(resultSet.getString("tenantId"))
-                    .dateOfReconciliation(resultSet.getInt("dateOfReconciliation"))
+                    .dateOfReconciliation(resultSet.getLong("dateOfReconciliation"))
                     .facilityId(resultSet.getString("facilityId"))
                     .calculatedCount(resultSet.getInt("calculatedCount"))
                     .commentsOnReconciliation(resultSet.getString("commentsOnReconciliation"))
